requirements.txt
flask~=3.0.0
wtforms~=3.1.1
werkzeug~=3.0.1
Pillow~=10.1.0
alembic~=1.12.1
SQLAlchemy~=2.0.23